fix: #247 - inference plugin should check nitro service available
