+++ electron/core/plugins/inference-plugin/index.ts
