+++ electron/core/plugins/inference-plugin/types/index.d.ts
@@ -1,0 +1,7 @@
+export {};
+
+declare global {
+  interface Window {
+    electronAPI?: any | undefined;
+  }
+}
